refactor wasm fixture gallery leaf1 module layout
