+++ scripts/wasm_fixture_gallery_v1/baseline.mjs
@@ -1,0 +1,3 @@
+export {
+  computeBaselineMatchV0,
+} from './main.mjs';
