+++ scripts/wasm_fixture_gallery_v1/delta_classifier.mjs
@@ -1,0 +1,5 @@
+export {
+  loadDeltaPolicyV1,
+  classifyBaselineCmpV1,
+  buildBaselineMetricsV1,
+} from './main.mjs';
