+++ scripts/wasm_fixture_gallery_v1/ondemand.mjs
@@ -1,0 +1,5 @@
+export {
+  mergeStoreRequestsV0,
+  normalizeStoreRequestFromEntryV0,
+  loadStoreRequestsV0,
+} from './main.mjs';
