+++ scripts/wasm_fixture_gallery_v1/report.mjs
@@ -1,0 +1,4 @@
+export {
+  buildConfigHashV0,
+  buildTypedArtifactsPlaceholderV0,
+} from './main.mjs';
