+++ scripts/wasm_fixture_gallery_v1/util_exec.mjs
@@ -1,0 +1,9 @@
+import { execFileSync } from 'node:child_process';
+
+export function runExecCheckedV0(command, args, options = {}) {
+  return execFileSync(command, args, {
+    encoding: 'utf8',
+    stdio: ['ignore', 'pipe', 'pipe'],
+    ...options,
+  });
+}
