+++ scripts/wasm_fixture_gallery_v1/wasm_bridge.mjs
@@ -1,0 +1,5 @@
+export {
+  readArtifactBytesV0,
+  readLogBytesV0,
+  entrypointSetOkV0,
+} from './main.mjs';
